pyproject.toml
[project]
name = "tdsproject2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.56.0",
    "beautifulsoup4>=4.14.2",
    "langgraph>=1.0.3",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "google-genai>=0.17.0",
    "jsonpatch>=1.33",
    "python-dotenv>=1.2.1",
    "pandas>=2.3.3",
    "fastapi>=0.121.3",
    "uvicorn>=0.38.0",
    "requests>=2.32.5",
    "ruff>=0.14.6",
    "langchain-openai>=1.1.0",
    "openai>=2.8.1",
    "tenacity>=9.1.2",
    "langchain-core>=1.1.0",
    "speechrecognition>=3.14.4",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "pytesseract>=0.3.13",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "matplotlib>=3.10.7",
    "pypdf2>=3.0.1",
    "geopy>=2.4.1",
    "networkx>=3.6",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.3",
    "langchain-ollama>=1.0.0",
    "html2text>=2025.4.15",
]
